feat : ChessService 테스트 구현
