+++ src/main/resources/templates/chessGame.html
@@ -196,6 +196,8 @@
 
 
 
+
+
 </script>
 
 <script>
